BankManagementSystem.py
The project name is Bank Management System. A Bank Management System keeps track of the individual account of the customer. With the help, of Python GUI
and My SQL, an application is created and all the information is stored in the backend using My SQL. It is also an offline application.